requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.13.0
beautifulsoup4==4.13.5
certifi==2026.4.22
charset-normalizer==3.4.7
click==8.3.3
colorama==0.4.6
fastapi==0.136.1
h11==0.16.0
idna==3.13
lxml==6.0.2
numpy==2.4.4
openai==2.11.0
pinecone[grpc]==7.0.2
psycopg2-binary==2.9.12
pydantic_core==2.46.4
pydantic==2.13.4
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
rank-bm25==0.2.2
requests==2.33.1
SQLAlchemy==2.0.49
starlette==1.0.0
trafilatura==2.0.0
typing_extensions==4.15.0
typing-inspection==0.4.2
urllib3==2.7.0
uvicorn==0.46.0
